Moved functions around + Updated CLI project file
