+++ Client/Client.cpp
@@ -4,7 +4,7 @@
 
 #include <iostream>
 #include <string>
-#include "..\Communication\Communication.h"
+#include "..\Communication\Create.h"
 #include "Client.h"
 
 
@@ -21,7 +21,25 @@
 //   6. In the future, to open this project again, go to File > Open > Project and select the .sln file
 
 
+Client::Client::Client()
+{
+	Data::IPV4Address localhost = { 127, 0, 0, 1 };
+	auto remote_host = Create::RemoteTcpServer(localhost);
+	this->sender = Create::TcpClient(localhost, std::move(remote_host));
+
+}
+
+Client::Client::Client(std::unique_ptr<Communicators::ISender> sender)
+	: sender(std::move(sender))
+{
+
+}
+
+inline void Client::Client::__setResponseParser(ResponseParser rp) {
+	this->response_parser = rp;
+}
+
 std::unique_ptr<Data::IServerResponse> Client::Client::SendCommand(std::unique_ptr<Data::IClientRequest> request)
 {
-    return std::unique_ptr<Data::IServerResponse>();
+    return Create::ServerResponse();
 }
