+++ Client/Client.h
@@ -8,7 +8,7 @@
 typedef std::function<std::unique_ptr<Data::IServerResponse>(std::vector<uint8_t>)> ResponseParser;
 
 namespace Client {
-	const ResponseParser default_parser = [](std::vector<uint8_t> response_serialization)
+	ResponseParser default_parser = [](std::vector<uint8_t> response_serialization)
 	{
 		return Create::ServerResponse(response_serialization);
 	};
@@ -25,24 +25,12 @@ namespace Client {
 		
 		ResponseParser response_parser;
 	public:
-		Client()
-		{
-			Data::IPV4Address localhost = { 127, 0, 0, 1 };
-			auto remote_host = Create::RemoteTcpServer(localhost);
-			this->sender = Create::TcpClient(localhost, std::move(remote_host));
-			
-		}
+		Client();
 
-		Client(std::unique_ptr<Communicators::ISender> sender)
-		: sender(std::move(sender))
-		{
+		Client(std::unique_ptr<Communicators::ISender> sender);;
 
-		};
+		void __setResponseParser(ResponseParser rp);
 
-		void __setResponseParser(ResponseParser rp) {
-			this->response_parser = rp;
-		}
-
-		std::unique_ptr<Data::IServerResponse> SendCommand(std::unique_ptr < Data::IClientRequest> request);
+		std::unique_ptr<Data::IServerResponse> SendCommand(std::unique_ptr <Data::IClientRequest> request);
 	};
 }
