+++ Client/Create.cpp
@@ -1,3 +1,4 @@
+#include "../Communication/Create.h"
 #include "Create.h"
 
 std::unique_ptr<Client::Client> Create::Client()
