+++ ClientCLI/ClientCLI.cpp
@@ -2,12 +2,11 @@
 //
 
 #include <iostream>
+#include <queue>
 #include "ClientCLI.h"
-#include "..\Client\Create.h"
-#include "..\Communication\Create.h"
-#include "..\Logger\Create.h"
-#include "..\Server\Create.h"
-#include "..\Tamagotchi\Create.h"
+
+
+
 
 void app(istream& input, ostream& output)
 {
